Scripts moved to lib folder
